Fix side-effect with BGPlay
