+++ src/App.vue
@@ -1,225 +1,30 @@
 <template>
   <q-layout view="hHh LpR fff" id="app" class="IHR_minimum-width bg-white">
-    <q-header elevated primary  class="IHR_minimum-width">
-      <q-toolbar class="q-py-sm q-px-lg row">
-        <div class="col-8 row no-wrap items-center">
-          <div>
-            <!--<q-btn flat @click="expandSidebar">-->
-              <!--<q-icon name="fas fa-bars" class="text-white" />-->
-            <!--</q-btn>-->
-          </div>
-          <q-item id="IHR_home-button">
-            <router-link :to="{name : 'home'}">
-              <q-btn round dense flat :ripple="false" no-caps size="22px">
-                <img src="@/assets/imgs/ihr_logo.svg" style="width: 45px;" />
-              </q-btn>
-            </router-link>
-          </q-item>
-          <network-search-bar/>
-
-          <div
-            class="IHR_menu-entries q-ml-xs q-gutter-md text-body2 text-weight-bold row items-center no-wrap"
-          >
-            <router-link
-              :key="item.entryName"
-              :to="{name : item.routeName}"
-              v-for="item in simple_menu"
-            >{{$t(item.entryName)}}</router-link>
-          </div>
-        </div>
-
-        <!--Log in /Log out stuff here-->
-      </q-toolbar>
-    </q-header>
+    <router-view name="header"/>
     <q-page-container class="IHR_minimum-width">
-      <router-view :show-sidebar="sidebarOpened" @sidebar-action="sidebarOpened = $event === true" />
+      <router-view/>
       <div id="IHR_last-element">&nbsp;</div>
     </q-page-container>
-    <q-footer
-      elevated
-      primary
-      class="IHR_minimum-width IHR_footer text-white text-center row items-streatch content-stretch no-wrap justiy-between"
-    >
-      <div class="col-2 IHR_copyright IHR_fsection q-pt-md">
-        <div class="text-weight-bold">
-          Internet Health Report
-        </div>
-        <div>
-          <router-link :to="{name : 'home'}">
-            <q-btn round dense flat :ripple="false" no-caps size="22px">
-              <img src="@/assets/imgs/ihr_logo.svg" style="width: 50px;" />
-            </q-btn>
-          </router-link>
-        </div>
-      </div>
-      <div class="IHR_fsection col-8 q-pt-md">
-        <div class="IHR_sitemap row justify-start">
-        <span class="col-3">
-          <ul>
-            <li class="text-weight-bold">
-                {{$t('footer.reportPages.title')}}
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'global_report'}"
-              >{{$t('footer.reportPages.global')}}</router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'networks', params:{ asn: 2497 }}"
-              >{{$t('footer.reportPages.network')}}</router-link>
-            </li>
-          </ul>
-        </span>
-        <span class="col-3">
-          <ul>
-            <li class="text-weight-bold">
-              <router-link
-                :to="{name : 'documentation'}" >
-                  {{$t('footer.documentation.title')}}
-              </router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'documentation'}"
-              >{{$t('footer.about.title')}}</router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'documentation', hash:'#AS_dependency'}"
-              >{{$t('footer.documentation.analysisModules')}}</router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'documentation', hash:'#REST_API' }"
-              >{{$t('footer.documentation.dataAccess')}}</router-link>
-            </li>
-          </ul>
-        </span>
-        <span class="col-3">
-          <ul>
-            <li class="text-weight-bold">
-              <router-link
-                :to="{name : 'contact'}" >
-                  {{$t('footer.about.contact')}}
-                </router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'documentation', hash:'#Acknowledgements'}"
-              >{{$t('footer.about.acknowledgments')}}</router-link>
-            </li>
-            <li>
-              <router-link
-                :to="{name : 'documentation', hash:'#Data_policy'}"
-              >{{$t('footer.about.datapolicy')}}</router-link>
-            </li>
-          </ul>
-        </span>
-        </div>
-        <div class="row items-center q-pt-lg">
-            <div class="col-2"><a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/4.0/"><img alt="Creative Commons License" style="border-width:0" src="https://i.creativecommons.org/l/by-nc-sa/4.0/88x31.png" /></a></div>
-            <div class="col-10 text-caption text-left q-pa-md">
-                <span xmlns:dct="http://purl.org/dc/terms/" href="http://purl.org/dc/dcmitype/Dataset" property="dct:title" rel="dct:type"> Internet Health Report</span> is licensed under a <a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/4.0/">Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License</a>. Permissions beyond the scope of this license may be available at <a xmlns:cc="http://creativecommons.org/ns#" href="mailto:ihr-admin@iij-ii.co.jp" rel="cc:morePermissions">ihr-admin@iij-ii.co.jp</a>.
-            </div>
-        </div>
-      </div>
-      <div class="IHR_external-links col-2 IHR_fsection">
-        <div class="row wrap justify-center q-pt-lg">
-          <a href="https://twitter.com/ihr_alerts" class="col-1" target="_blank">
-            <q-icon name="fab fa-twitter-square" />
-          </a>
-          <div class="col-1 q-pa-md"></div>
-          <a href="https://github.com/InternetHealthReport" class="col-1" target="_blank">
-            <q-icon name="fab fa-github-square" />
-          </a>
-          <div class="col-1 q-pa-md"></div>
-          <router-link :to="{name : 'contact'}" class="col-1" target="_blank">
-            <q-icon name="fas fa-envelope-square" />
-          </router-link>
-        </div>
-      </div>
-    </q-footer>
+    <router-view name="footer"/>
   </q-layout>
 </template>
 <script>
 import languages from "quasar/lang/index.json";
-import LocaleSelector from "@/locales/LocaleSelector";
-import NetworkSearchBar from "@/components/search_bar/NetworkSearchBar";
 import routerBase from "@/router";
-import LoginForm from "@/components/forms/LoginForm";
 
 // subset of router, see router.js
-
-const simple_menu = [
-  {
-    entryName: "header.home",
-    routeName: "home"
-  },
-  {
-    entryName: "header.globalReport",
-    routeName: "global_report"
-  },
-  {
-    entryName: "header.documentation",
-    routeName: "documentation"
-  },
-  {
-    entryName: "header.API",
-    routeName: "api"
-  },
-  {
-    entryName: "header.contact",
-    routeName: "contact"
-  }
-];
-
 export default {
   name: "Default",
   components: {
-    LocaleSelector,
-    NetworkSearchBar,
-    LoginForm
   },
   data() {
     return {
       text: "",
-      simple_menu: simple_menu,
-      sidebarOpened: false,
-      loginError: false
     };
   },
   mounted(){ 
     document.title = 'Internet Health Report';
   },
-  methods: {
-    expandSidebar() {
-      this.sidebarOpened = !this.sidebarOpened;
-    },
-    login(email, password) {
-      if (
-        this.$ihrStyle.validateEmail(email) &&
-        this.$ihrStyle.validatePassword(password)
-      ) {
-        this.$ihr_api.userLogin(
-          email,
-          password,
-          () => {},
-          () => {
-            this.loginError = true;
-          }
-        );
-      }
-    },
-    logout() {
-      this.$ihr_api.userLogout();
-    }
-  },
-  watch: {
-    '$route' (to, from) {
-      document.title = to.meta.title || 'IHR'
-    }
-  }
 };
 </script>
 <style lang="stylus">
